+++ CHANGELOG.md
@@ -2,3 +2,4 @@
 
 * Fixed Occasional crashes inside of ChunkMixin#onAddingBlock
 * Added environment transfer / adjacent transfer to the Mek Heat Trait.
+* QoL Being able to change recipe ids (create and copy)